Users-service/pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.14"
fastapi = {extras = ["standart"], version = "^0.128.0"}
uvicorn = {extras = ["standart"], version = "^0.38.0"}
orjson = "^3.11.7"
pydantic = "^2.12.5"
pydantic-settings = "^2.12.0"
python-multipart = "^0.0.22"
loguru = "^0.7.3"
sqlalchemy = "^2.0.46"
asyncpg = "^0.31.0"
sqlalchemy-utc = "^0.14.0"
alembic = "^1.18.3"
prometheus-fastapi-instrumentator = "^7.1.0"
pyjwt = "^2.11.0"
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"
cryptography = "^46.0.5"
redis = "^7.2.0"
sqladmin = "^0.23.0"
email-validator = "^2.3.0"
bcrypt = "^5.0.0"

[dependency-groups]
dev = [
    "black (>=25.12.0,<26.0.0)"
]
